analysis_pipeline.py
"""
analysis_pipeline.py
====================
層級三：農業統計數據前處理、特徵工程與訓練集輸出。

職責範圍：
  層級三前半（對話四）— 農業統計數據整合與前處理：
    - 從 moa_opendata 資料庫讀取 11 本農業統計數據集
    - 確認各資料集結構、可用欄位與時間範圍
    - 設計並執行大型資料集（3 本 20~80 萬筆）的彙總策略：
        時間維度壓縮（年均、最近 N 年）、作物分類彙總、統計方式選擇
    - 將明細資料壓縮為每鄉鎮一筆的特徵向量
    - 與 GIS 量化輸出檔案 JOIN（key：townid / countyid）
    - 缺值處理、標準化 / 正規化、類別變數編碼

  層級三後半（對話五）— 特徵工程與訓練集輸出：
    - 衍生特徵設計（比例、差值、時間趨勢等）
    - EDA（分布圖、相關性矩陣）
    - 輸出訓練集（.csv 或 .parquet）
    - 整合入口完成（main_pipeline.py，含 argparse 排程支援）

設計原則：
  - 此模組只負責「處理」，不持有任何資料庫連線；conn 由呼叫端傳入
  - GIS 量化結果以檔案路徑（output/*.csv / *.parquet）傳入，不重新查詢資料庫
  - 所有前處理步驟產出 preprocessed_metadata dict，記錄清理規則與統計摘要，
    方便後續重現與稽核
  - 輸出訓練集命名規範：
      output/GIS_processed_{layer}_{YYYYMMDD}.csv
      output/GIS_processed_{layer}_{YYYYMMDD}.parquet

輸入來源：
  - moa_opendata 資料庫（農業統計數據，對話四建立連線）
  - output/GIS_processed_002_farmland_{date}.csv / .parquet（WMS 量化結果）
  - output/GIS_processed_003_suitability_{date}.csv / .parquet

輸出目標：
  - output/ 下的訓練集 .csv / .parquet，供機器學習模型直接使用

TODO（Phase 4 — 對話四）：
  - [ ] 實作 load_agricultural_statistics(conn, dataset_name)：
        從 moa_opendata 讀取指定農業統計數據集，回傳 DataFrame
  - [ ] 實作 summarize_large_dataset(df, group_keys, agg_strategy)：
        針對 20~80 萬筆的大型資料集執行彙總，
        將明細資料壓縮為每鄉鎮一筆的特徵向量；
        agg_strategy 設計待與 Claude 協作確認（對話四）
  - [ ] 實作 join_gis_and_statistics(gis_df, stat_df, join_key)：
        以 townid / countyid 為 key，合併 GIS 量化結果與農業統計特徵向量
  - [ ] 實作 handle_missing_values(df, strategy_map)：
        依欄位類型套用不同缺值處理策略（填中位數、填 0、刪除欄位等）
  - [ ] 實作 encode_categorical(df, columns)：
        對類別變數執行 One-Hot 或 Label Encoding
  - [ ] 實作 normalize_features(df, columns, method)：
        標準化 / 正規化數值特徵（StandardScaler / MinMaxScaler）

TODO（Phase 5 — 對話五）：
  - [ ] 實作 engineer_features(df)：
        衍生特徵設計，例如：
        - 各類別土地覆蓋率的比例特徵（A類面積 / 總面積）
        - 跨年差值（今年產量 - 去年產量）
        - 時間趨勢（近 N 年線性斜率）
  - [ ] 實作 run_eda(df, output_dir)：
        輸出分布圖（histogram）與相關性矩陣（heatmap）至 output/eda/
  - [ ] 實作 export_training_set(df, layer_name, format)：
        將最終訓練集輸出為 .csv 或 .parquet，
        檔名格式：GIS_processed_{layer}_{YYYYMMDD}.{ext}

相依模組（預期）：
  pandas, numpy
  sklearn.preprocessing（標準化、編碼，對話五引入）
  logs_handle（統一 logger）
"""

import pandas as pd
from typing import Tuple
from logs_handle import logger


def analyze_and_clean_dataframe(df: pd.DataFrame) -> Tuple[pd.DataFrame, dict]:
    """
    對 DataFrame 進行基礎清理，移除全空行，並產出前處理 metadata 摘要。

    此為最小可用版本，目前僅執行「刪除所有欄位皆為空的行」一項規則。
    完整的清理流水線將在對話四依實際資料集結構擴充。

    Args:
        df (pd.DataFrame): 原始輸入 DataFrame，通常來自農業統計數據集或
            GIS 量化結果的中間處理階段。

    Returns:
        Tuple[pd.DataFrame, dict]:
          - cleaned_df：清理後的 DataFrame，index 未重設
          - preprocessed_metadata：前處理摘要，結構如下：
              {
                "cleaning_rules": str,           # 本次套用的清理規則說明
                "rows_deleted": int,             # 刪除的行數
                "data_types_before_cleaning": {col: dtype_str, ...},
                "data_types_after_cleaning":  {col: dtype_str, ...}
              }

    Notes:
        - dropna(how='all') 只刪除「所有欄位皆為 NaN」的行，
          部分欄位為 NaN 的行不受影響，需在對話四依欄位語意另行處理
        - preprocessed_metadata 設計為可序列化（dict of str），
          方便後續以 file_utils.save_json_data() 落地留存

    TODO（Phase 4）：
        - [ ] 依各資料集實際結構擴充清理規則：
              - 特定欄位的缺值填補（中位數、眾數、0）
              - 數字字串轉數值型態（例如產量欄位含千分位逗號）
              - 移除重複行（dropna 後再 drop_duplicates）
              - 異常值偵測（IQR 或 z-score）
        - [ ] 將 cleaning_rules 從固定字串改為 list，
              每新增一條規則就 append，方便 metadata 精確記錄實際執行內容
        - [ ] 考慮加入 reset_index(drop=True)，避免下游 iloc / merge 行為不一致
    """
    initial_rows = df.shape[0]
    cleaned_df = df.dropna(how='all')
    rows_deleted = initial_rows - cleaned_df.shape[0]

    preprocessed_metadata = {
        "cleaning_rules": "刪除所有欄位皆為空的行",
        "rows_deleted": rows_deleted,
        "data_types_before_cleaning": {col: str(df[col].dtype) for col in df.columns},
        "data_types_after_cleaning": {col: str(cleaned_df[col].dtype) for col in cleaned_df.columns}
    }

    logger.notice(
        f"資料清理完成。原始行數：{initial_rows}，"
        f"清理後行數：{cleaned_df.shape[0]}，刪除行數：{rows_deleted}"
    )

    return cleaned_df, preprocessed_metadata


# TODO（Phase 4）：實作 load_agricultural_statistics(conn, dataset_name) -> pd.DataFrame
# TODO（Phase 4）：實作 summarize_large_dataset(df, group_keys, agg_strategy) -> pd.DataFrame
# TODO（Phase 4）：實作 join_gis_and_statistics(gis_df, stat_df, join_key) -> pd.DataFrame
# TODO（Phase 4）：實作 handle_missing_values(df, strategy_map) -> pd.DataFrame
# TODO（Phase 4）：實作 encode_categorical(df, columns) -> pd.DataFrame
# TODO（Phase 4）：實作 normalize_features(df, columns, method) -> pd.DataFrame
# TODO（Phase 5）：實作 engineer_features(df) -> pd.DataFrame
# TODO（Phase 5）：實作 run_eda(df, output_dir) -> None
# TODO（Phase 5）：實作 export_training_set(df, layer_name, format) -> str
