requirements.txt
bs4==0.0.2
dotenv==0.9.9
geopandas==1.0.1
matplotlib==3.9.4
mss==10.1.0
opencv-python==4.12.0.88
pip-chill==1.0.3
psycopg2-binary==2.9.11
pyautogui==0.9.54
pynput==1.8.1
pysocks==1.7.1
pytesseract==0.3.13
rasterio==1.4.3
requests==2.32.5
scikit-learn==1.6.1
selenium==4.36.0
torch==2.8.0
tqdm==4.67.1
vincenty==0.1.4
